Fix Arrow configure in Conda envs
If $CONDA_PREFIX is set Arrow will try to find its dependencies only in
that conda environment. However we want to let it find them either in
the EB installed folders or fallback to the bundled sources.
So pretend we are not in a conda environment by unsetting the variable.